Finalizações

+++ src/main/java/br/gov/sp/fatec/bibliotecaOnline/Services/LivroService.java
@@ -4,9 +4,9 @@ import br.gov.sp.fatec.bibliotecaOnline.Entities.Livro;
 import java.util.List;
 
 public interface LivroService {
-    public Livro createLivro(String nomeTitulo, String editora, Float preco, String autor, String sessao, List<String> categorias);
+    public Livro createLivro(String nomeTitulo, String editora, Double preco, String autor, String sessao, List<String> categorias);
     public Livro readLivro(Integer id);
-    public Livro updateLivro(Integer id, String nomeTitulo, String editora, Float preco);
+    public Livro updateLivro(Integer id, String nomeTitulo, String editora, Double preco);
     public Boolean deleteLivro(Integer id);
     public List<Livro> getByAutorAndSessao(Integer idAutor, Integer idSessao);
 }
